removed the 'private' method in favor of a more functional approach
