+++ lib/set.js
@@ -12,6 +12,7 @@
 
 var EventEmitter = require('events')
 var Migration = require('./migration')
+var migrate = require('./migrate')
 var fs = require('fs')
 var inherits = require('inherits')
 
@@ -133,63 +134,6 @@ MigrationSet.prototype.migrate = function (direction, migrationName, fn) {
     } else {
       this.pos = obj.pos
     }
-    this._migrate(direction, migrationName, fn)
+    migrate(this, direction, migrationName, fn)
   })
-}
-
-/**
- * Get index of given migration in list of migrations
- *
- * @api private
- */
-
-function positionOfMigration (migrations, filename) {
-  for (var i = 0; i < migrations.length; ++i) {
-    if (migrations[i].title === filename) return i
-  }
-  return -1
-}
-
-/**
- * Perform migration.
- *
- * @api private
- */
-
-MigrationSet.prototype._migrate = function (direction, migrationName, fn) {
-  var migrations
-  var migrationPos
-
-  if (!migrationName) {
-    migrationPos = direction === 'up' ? this.migrations.length : 0
-  } else if ((migrationPos = positionOfMigration(this.migrations, migrationName)) === -1) {
-    return fn(new Error('Could not find migration: ' + migrationName))
-  }
-
-  switch (direction) {
-    case 'up':
-      migrations = this.migrations.slice(this.pos, migrationPos + 1)
-      break
-    case 'down':
-      migrations = this.migrations.slice(migrationPos, this.pos).reverse()
-      break
-  }
-
-  var next = (migration) => {
-    if (!migration) return fn(null)
-
-    this.emit('migration', migration, direction)
-    migration[direction]((err) => {
-      if (err) return fn(err)
-
-      this.pos += (direction === 'up' ? 1 : -1)
-      this.save((err) => {
-        if (err) return fn(err)
-
-        next(migrations.shift())
-      })
-    })
-  }
-
-  next(migrations.shift())
 }
